price as int

+++ src/main/java/com/jakubadamek/robotemil/entities/PriceAndOrder.java
@@ -7,8 +7,8 @@ import java.io.Serializable;
  */
 public class PriceAndOrder implements Serializable {
 	private static final long serialVersionUID = 1482674536827012069L;
-	/** price */
-	public Double price;
+	/** price times 100 */
+	public int price;
 	/** order on the web site */
 	public int order;
